debuginfo: extract adt.rs
